pyproject.toml
[tool.poetry]
name = "reservoirdb"
version = "0.1.0"
description = ""
authors = [ "Ruan Pearce-Authers <ruan@reservoirdb.com>" ]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.3"
dacite = "^1.6.0"
pyarrow = "^3.0.0"
pandas = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.30.0"
mypy = "^0.800"
pytest-asyncio = "^0.14.0"
python-dotenv = "^0.15.0"

[build-system]
requires = [ "poetry-core >= 1.0.0" ]
build-backend = "poetry.core.masonry.api"
